// src/app/shared/models/branch.ts
export class Branch {
  id?: number ;
  name_en?: string = '';
  name_ar?: string = '';
  address1?: string = '';
  address2?: string = '';
  city_id?: number ;
  phone?: string = '';
  branch_category?: string = '';
  branch_capacity?: number ;
  startup_date?: string = '';
  email?: string = '';
  lat?: number ;
  lng?: number ;
  location_type?: string = '';
}