// src/app/shared/models/signup.ts
export class Signup {
  card_number?: number;
  client_id?: number;
  cvv?: string;
  date_month?: string;
  date_year?: string;
  default_branch?: number;
  dob?: string;
  email?: string;
  fcm_token?: string;
  first_name?: string;
  gender?: string;
  last_name?: string;
  mobile?: string;
  national_id?: string;
  password?: string;
  referral_code?:string;
  preferred_language?: string;
  promotional_code?: string;
  subscription_day?: string;
  product_id?: number;
  member_id?: number;
  parent_mobile?: string;
  parent_name?: string;
}